+++ src/pip/_internal/metadata/pkg_resources.py
@@ -16,7 +16,8 @@ from pip._vendor.packaging.utils import NormalizedName, canonicalize_name
 from pip._vendor.packaging.version import Version
 from pip._vendor.packaging.version import parse as parse_version
 
-from pip._internal.exceptions import InvalidWheel, NoneMetadataError, UnsupportedWheel
+from pip._internal.exceptions import NoneMetadataError
+from pip._internal.exceptions.wheel import InvalidWheel, UnsupportedWheel
 from pip._internal.utils.egg_link import egg_link_path_from_location
 from pip._internal.utils.misc import display_path, normalize_path
 from pip._internal.utils.wheel import parse_wheel, read_wheel_metadata_file
