+++ tests/unit/test_utils_wheel.py
@@ -8,7 +8,7 @@ from zipfile import ZipFile
 
 import pytest
 
-from pip._internal.exceptions import UnsupportedWheel
+from pip._internal.exceptions.wheel import UnsupportedWheel
 from pip._internal.utils import wheel
 
 from tests.lib import TestData
