+++ tests/unit/test_wheel.py
@@ -198,7 +198,7 @@ def test_wheel_root_is_purelib(text: str, expected: bool) -> None:
 
 
 def test_dist_from_broken_wheel_fails(data: TestData) -> None:
-    from pip._internal.exceptions import InvalidWheel
+    from pip._internal.exceptions.wheel import InvalidWheel
     from pip._internal.metadata import FilesystemWheel, get_wheel_distribution
 
     package = data.packages.joinpath("corruptwheel-1.0-py2.py3-none-any.whl")
